Redis works...
Checking the database with the redis-cli shows that everything is
working as intended, and is being stored... but unfortunately I still
don't have a handle on the async thing. The state keeps returning as
"true" as opposed to an object with all the necessary properties.
Consequently, the client can't do anything with the non-existent
information.